Adiciona imagens as molduras
