+++ index.html
@@ -2,9 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Sun's Building</title>
+    <link rel="icon" type="image/x-icon" href="/Suns-building/src/assets/favicon.ico" />
   </head>
   <body>
     <div id="root"></div>
