+++ src/App.tsx
@@ -1,6 +1,5 @@
 import Header from "./components/Header";
 import Footer from "./components/Footer";
-import Principal from "./components/Principal";
 import Autenticar from "./pages/Autenticar";
 import Login from "./pages/Login";
 
@@ -24,6 +23,7 @@ export default function App() {
       <CardSmartCities></CardSmartCities>
       <CardEnergiaRenovaveis></CardEnergiaRenovaveis>
       <CardIcones></CardIcones>
+      <Footer />
     </>
   )
 }
