+++ src/components/CardEdificio.tsx
@@ -1,17 +1,14 @@
-import React from 'react'
-
-//CSS
 import '../styles/CardEdificio.css'
 
-//img
 import backGrey from '../assets/fb51434f75f4513fd43833fba4c4f2d6.jpg'
+import { buscarImagem } from './buscarImagem'
 
 
 const CardEdificio = () => {
 
     return (
     <div className='container-cardedificio'>
-        <img src={backGrey} alt="" />
+        <img src={buscarImagem('centralBuilding')} alt="" />
     </div> 
   )
 }
