+++ src/components/CardEnergiaRenovaveis.tsx
@@ -1,10 +1,5 @@
-import React from 'react'
-
-//Estilo
 import '../styles/CardEnergiaRenovaveis.css'
-
-//img 
-import backGrey from '../assets/fb51434f75f4513fd43833fba4c4f2d6.jpg'
+import { buscarImagem } from './buscarImagem'
 
 const CardEnergiaRenovaveis = () => {
   return (
@@ -27,7 +22,7 @@ const CardEnergiaRenovaveis = () => {
 
         
         <div className='container-img-cardenergiarenovaveis'>
-            <img src={backGrey} alt="" />
+            <img className='imagens-pequenas' src={buscarImagem('geothermalEnergy')} alt="" />
         </div>
