--- src/components/Principal.tsx
@@ -1,9 +1,0 @@
-function Principal() {
-  return (
-    <>
-      <h1>Aqui fica o conteudo principal, olhem ""../src/index.css" pra entender como sera feito o estilo</h1>
-    </>
-  )
-}
-
-export default Principal;
