+++ src/styles/CardBioArquitetura.css
@@ -16,3 +16,8 @@
 .lista-bioarquitetura li{
     margin-bottom: 30px;
 }
+
+.imagens-pequenas {
+    height: 417;
+    width: 626;
+}
