+++ src/styles/CardEnergiaRenovaveis.css
@@ -17,3 +17,8 @@
 .lista-energiarenovaveis li{
     margin-bottom: 30px;
 }
+
+.imagens-pequenas {
+    height: 417px;
+    width: 626px;
+}
