+++ src/styles/CardSmartCities.css
@@ -17,3 +17,8 @@
 .lista-smartcities li{
     margin-bottom: 30px;
 }
+
+.imagens-pequenas {
+    height: 417px;
+    width: 626px;
+}
