+++ src/styles/CardsTopicos.css
@@ -21,4 +21,5 @@
     max-width: 100%;
     height: 450px;
     margin-top: 30px;
+    width: 600px;
 }
