+++ src/styles/Header.css
@@ -47,5 +47,8 @@
 
 .header-right ul li{
     list-style: none;
+}
 
+li {
+    list-style-type: none;   
 }
